added gameString attribute
now the entire game could be saved as a gameString. This could mean more computation in game but also helps to save games to a more readable format at the end.
Another alternative is to calculate the gameString inside the save method. This could be clean in terms of efficiency but would increase code.